Add files via upload
This chart shows the average duration of Physical Therapy session (in minutes) provided to patients in a hospital (phantom) based on either initial outcome measure score (AMPAC score), or the specified service line patient was admitted to. All data used in this project were artificially generated using excel formulas. Most of the AMPAC scores trending negative are not typical of general trends (it happened due to the level of difficulty generating incremental random numbers for a variety of ID numbers). If real values were used then most if not all of the trends will at least be positive.  Note(as I continue to work on this project, adjustments will be made to make all data look as realistic as possible).